web/backend/requirements.txt
# ===========================================
# MiroShark Backend Dependencies
# ===========================================
# Python 3.11+ required
# Install: pip install -r requirements.txt
# ===========================================

# ============= Core Framework =============
flask>=3.0.0
flask-cors>=6.0.0
flask-compress>=1.17

# ============= LLM Related =============
# OpenAI SDK (unified LLM calls using OpenAI format)
openai>=1.0.0

# ============= Neo4j =============
neo4j>=5.15.0

# ============= HTTP Requests =============
# For Ollama embedding API
requests>=2.31.0

# ============= Wonderwall Social Media Simulation (Wonderwall fork) =============
# Wonderwall (wonderwall package) is bundled directly in backend/wonderwall/
# No separate install needed — camel-oasis is no longer a pip dependency
camel-ai==0.2.78

# ============= File Processing =============
PyMuPDF>=1.24.0
# Encoding detection (supports non-UTF-8 encoded text files)
charset-normalizer>=3.0.0
chardet>=5.0.0

# ============= Utility Libraries =============
# Environment variable loading
python-dotenv>=1.0.0

# Data validation
pydantic>=2.0.0

# ============= Browser Push Notifications =============
# Required for Web Push (VAPID key generation + push dispatch)
pywebpush>=2.0.0

# ============= Game Theory (optional — report_agent.analyze_equilibrium) =============
# Enables the Nash equilibrium tool on the ReACT report agent. Silent no-op if absent.
nashpy>=0.0.41
